Changed ordering value from long to double in opt_order.cc, for some problems in IPC18 competition, e.g. civ domain, long was too small.  This value is used to decide on which order to set the variables.  We found this problem while checking a out of memory issue in symbolic while setting up its TRs.  Note: This change will not fix all the memory out issues, as this is dependent on the conditional effects being grouped close together, this is somewhat stocastic, it is not the direct objective of the ordering function.  To fully fix, it would require a new ordering method based on controlling the max size of TRs.

+++ src/search/symbolic/opt_order.cc
@@ -44,13 +44,13 @@ void InfluenceGraph::compute_gamer_ordering(std::vector <int> &var_order) {
 
 
 void InfluenceGraph::get_ordering(vector <int> &ordering) const {
-    long value_optimization_function = optimize_variable_ordering_gamer(ordering, 50000);
+    double value_optimization_function = optimize_variable_ordering_gamer(ordering, 50000);
     DEBUG_MSG(cout << "Value: " << value_optimization_function << endl;);
 
     for (int counter = 0; counter < 20; counter++) {
         vector <int> new_order;
         randomize(ordering, new_order); //Copy the order randomly
-        long new_value = optimize_variable_ordering_gamer(new_order, 50000);
+        double new_value = optimize_variable_ordering_gamer(new_order, 50000);
 
         if (new_value < value_optimization_function) {
             value_optimization_function = new_value;
@@ -86,11 +86,11 @@ void InfluenceGraph::randomize(vector <int> &ordering, vector<int> &new_order) {
 
 
 
-long InfluenceGraph::optimize_variable_ordering_gamer(vector <int> &order,
+double InfluenceGraph::optimize_variable_ordering_gamer(vector <int> &order,
                                                       int iterations) const {
-    long totalDistance = compute_function(order);
+    double totalDistance = compute_function(order);
 
-    long oldTotalDistance = totalDistance;
+    double oldTotalDistance = totalDistance;
     //Repeat iterations times
     for (int counter = 0; counter < iterations; counter++) {
         //Swap variable
@@ -136,8 +136,8 @@ long InfluenceGraph::optimize_variable_ordering_gamer(vector <int> &order,
 
 
 
-long InfluenceGraph::compute_function(const std::vector <int> &order) const {
-    long totalDistance = 0;
+double InfluenceGraph::compute_function(const std::vector <int> &order) const {
+    double totalDistance = 0;
     for (size_t i = 0; i < order.size() - 1; i++) {
         for (size_t j = i + 1; j < order.size(); j++) {
             if (influence(order[i], order[j])) {
@@ -162,9 +162,9 @@ void InfluenceGraph::optimize_variable_ordering_gamer(vector <int> &order,
                                                       vector <int> &partition_begin,
                                                       vector <int> &partition_sizes,
                                                       int iterations) const {
-    long totalDistance = compute_function(order);
+    double totalDistance = compute_function(order);
 
-    long oldTotalDistance = totalDistance;
+    double oldTotalDistance = totalDistance;
     //Repeat iterations times
     for (int counter = 0; counter < iterations; counter++) {
         //Swap variable
